requirements.txt
google-genai>=1.0.0
playwright>=1.52.0
PySide6>=6.8.0
python-docx>=1.1.0
tenacity>=8.2.0
bing-image-urls>=0.1.5
Pillow>=10.0.0
